special dashes guide

+++ Guidebook/Pages/AboutGodSouls.cs
@@ -4,7 +4,7 @@ namespace ArcaneOdyssey.Guidebook.Pages
 {
 	public class AboutGodSouls : GuidebookPage
 	{
-		public override int PageNum => After<ForgingBronze>();
+		public override ushort PageNum => After<ForgingBronze>();
 
 		public override bool MetConditions(Player player) => (player.ArcaneOdyssey()?.Souls.Count ?? 1) > 1;
 	}
