+++ Guidebook/Pages/MagicTypes.cs
@@ -6,6 +6,6 @@ namespace ArcaneOdyssey.Guidebook.Pages
 	public class MagicTypes : GuidebookPage
 	{
 		public override bool MetConditions(Player player) => player.HasTypeInInventory<MagicType>();
-		public override int PageNum => After<GettingStarted>();
+		public override ushort PageNum => After<GettingStarted>();
 	}
 }
