+++ Guidebook/Pages/Mutating.cs
@@ -5,7 +5,7 @@ namespace ArcaneOdyssey.Guidebook.Pages
 {
 	public class Mutating : GuidebookPage
 	{
-		public override int PageNum => After<StrengthWeapons>();
+		public override ushort PageNum => After<StrengthWeapons>();
 
 		public override bool MetConditions(Player player) => player.HasTypeInInventory<HecateShard>();
 	}
