+++ Guidebook/Pages/Relics.cs
@@ -6,6 +6,6 @@ namespace ArcaneOdyssey.Guidebook.Pages
 	public class Relics : GuidebookPage
 	{
 		public override bool MetConditions(Player player) => player.HasTypeInInventory<SpiritEnergy>();
-		public override int PageNum => After<FightingStyles>();
+		public override ushort PageNum => After<FightingStyles>();
 	}
 }
