+++ Guidebook/Pages/StrengthWeapons.cs
@@ -4,7 +4,7 @@ namespace ArcaneOdyssey.Guidebook.Pages
 {
 	public class StrengthWeapons : GuidebookPage
 	{
-		public override int PageNum => After<AboutGodSouls>();
+		public override ushort PageNum => After<AboutGodSouls>();
 
 		public override bool MetConditions(Player player) => player.HasItemInInventory(e => ArcaneOdysseyMod.Sets.weaponType[e.type] == WeaponType.Strength);
 	}
