+++ Guidebook/Pages/UsingScrolls.cs
@@ -5,7 +5,7 @@ namespace ArcaneOdyssey.Guidebook.Pages
 {
 	public class UsingScrolls : GuidebookPage
 	{
-		public override int PageNum => After<ObtainingScrolls>();
+		public override ushort PageNum => After<ObtainingScrolls>();
 
 		public override bool MetConditions(Player player) => player.HasTypeInInventory<Scroll>();
 	}
