+++ Guidebook/Pages/Weapons.cs
@@ -5,7 +5,7 @@ namespace ArcaneOdyssey.Guidebook.Pages
 {
 	public class Weapons : GuidebookPage
 	{
-		public override int PageNum => After<ForgingBronze>();
+		public override ushort PageNum => After<ForgingBronze>();
 
 		public override bool MetConditions(Player player) => player.HasTypeInInventory<Weapon>(out var weap) && weap.Ability.HasValue;
 	}
